requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.17.0
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
certifi==2023.5.7
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
iso8601==1.1.0
magic-filter==1.0.9
multidict==6.0.4
pypika-tortoise==0.1.6
pytz==2023.3
tortoise-orm==0.19.3
typing_extensions==4.6.3
yarl==1.9.2
